refactor: inline remark helper into plugin package
