+++ .gitignore
@@ -40,6 +40,7 @@ build/Release
 # Dependency directories
 node_modules/
 jspm_packages/
+packages/docusaurus-plugin-smartlinker/dist/
 
 # npm lockfile (workspace is managed via pnpm)
 package-lock.json
