+++ examples/site/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@examples/site",
   "private": true,
-  
+
   "scripts": {
     "dev": "docusaurus start",
     "build": "docusaurus build",
@@ -10,7 +10,7 @@
   "dependencies": {
     "@docusaurus/core": "^3.0.0",
     "@docusaurus/preset-classic": "^3.0.0",
-    "docusaurus-plugin-smartlinker": "workspace:*",
+    "docusaurus-plugin-smartlinker": "file:../../packages/docusaurus-plugin-smartlinker",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   }
