+++ package.json
@@ -13,7 +13,7 @@
   },
   "scripts": {
     "build": "npm run build --workspace docusaurus-plugin-smartlinker",
-    "test": "npm run test --workspace docusaurus-plugin-smartlinker --workspace @internal/remark-smartlinker",
+    "test": "npm run test --workspace docusaurus-plugin-smartlinker",
     "lint": "echo \"@(optional) add eslint later\"",
     "site:dev": "npm run dev --workspace @examples/site",
     "site:build": "npm run build --workspace @examples/site",
