+++ packages/docusaurus-plugin-smartlinker/tests/remark/smoke.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import plugin from '../src/index.js';
+import plugin from '../../src/remark/index.js';
 
 describe('remark plugin skeleton', () => {
   it('exports a function', () => {
