+++ packages/docusaurus-plugin-smartlinker/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "../../tsconfig.base.json",
-  "include": ["src"],
+  "include": ["src/**/*.ts"],
+  "exclude": ["src/remark/**/*"],
   "compilerOptions": {
     "outDir": "./dist"
   }
