+++ packages/docusaurus-plugin-smartlinker/tsconfig.remark.cjs.json
@@ -1,0 +1,19 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "include": ["src/remark/**/*.ts"],
+  "compilerOptions": {
+    "outDir": "./dist/remark-cjs",
+    "rootDir": "./src/remark",
+    "baseUrl": "./",
+    "module": "CommonJS",
+    "moduleResolution": "node",
+    "paths": {
+      "docusaurus-plugin-smartlinker": [
+        "./dist/index.d.ts"
+      ],
+      "docusaurus-plugin-smartlinker/*": [
+        "./dist/*"
+      ]
+    }
+  }
+}
