--- packages/remark-smartlinker/tsconfig.cjs.json
@@ -1,12 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "CommonJS",
-    "moduleResolution": "node",
-    "outDir": "./dist-cjs",
-    "declaration": false,
-    "declarationMap": false,
-    "emitDeclarationOnly": false,
-    "sourceMap": false
-  }
-}
