--- packages/remark-smartlinker/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "include": ["src"],
-  "compilerOptions": {
-    "outDir": "./dist",
-    "baseUrl": "./",
-    "paths": {
-      "docusaurus-plugin-smartlinker": [
-        "../docusaurus-plugin-smartlinker/dist/index.d.ts",
-        "../docusaurus-plugin-smartlinker/src/index.ts"
-      ],
-      "docusaurus-plugin-smartlinker/*": [
-        "../docusaurus-plugin-smartlinker/dist/*",
-        "../docusaurus-plugin-smartlinker/src/*"
-      ]
-    }
-  }
-}
